Upgrade MySQL jdbc library. Fix import process (Import cause OOM error on large data) Show in task bar overall progress in windows 7 task bar.

+++ trunk/app/src/java/org/xblackcat/rojac/util/SynchronizationUtils.java
@@ -17,11 +17,11 @@ public final class SynchronizationUtils {
             return "";
         }
 
-        int size = e.getValue();
+        long size = e.getValue();
         String factors = " KMGTE";
         int idx = -1;
 
-        int lastSize;
+        long lastSize;
         do {
             lastSize = size;
             size >>= 10;
